Split out the classes/interfaces from the single file into their own files
