+++ src/GovUk.Education.ExploreEducationStatistics.Analytics.Consumer/Services/Workflow/IProcessRequestFilesWorkflow.cs
@@ -1,0 +1,6 @@
+﻿namespace GovUk.Education.ExploreEducationStatistics.Analytics.Consumer.Services.Workflow;
+
+public interface IProcessRequestFilesWorkflow
+{
+    Task Process(IWorkflowActor actor);
+}
