+++ src/GovUk.Education.ExploreEducationStatistics.Analytics.Consumer/Services/Workflow/ProcessRequestFilesWorkflow.cs
@@ -5,11 +5,6 @@ using Microsoft.Extensions.Logging;
 
 namespace GovUk.Education.ExploreEducationStatistics.Analytics.Consumer.Services.Workflow;
 
-public interface IProcessRequestFilesWorkflow
-{
-    Task Process(IWorkflowActor actor);
-}
-
 /// <summary>
 /// This class represents a common workflow that a number of IRequestFileProcessor
 /// implementations use in order to generate their Parquet report files.
@@ -230,107 +225,3 @@ public class ProcessRequestFilesWorkflow(
         });
     }
 }
-
-/// <summary>
-/// This interface represents a component that interacts with a file store
-/// comprising directories and files.
-///
-/// It declares common filesystem interactions, mostly modelled from equivalent
-/// calls from <see cref="Directory"/> and <see cref="File"/>.
-/// </summary>
-public interface IFileAccessor
-{
-    bool DirectoryExists(string directory);
-    
-    void CreateDirectory(string directory);
-    
-    void DeleteDirectory(string directory);
-    
-    IList<string> ListFiles(string directory);
-    
-    void Move(string sourcePath, string destinationPath);
-}
-
-/// <summary>
-/// This default implementation of <see cref="IFileAccessor"/> interacts
-/// directly with a standard filesystem, using <see cref="Directory"/>
-/// and <see cref="File"/> to perform the work.
-/// </summary>
-internal class FilesystemFileAccessor : IFileAccessor
-{
-    public bool DirectoryExists(string directory)
-    {
-        return Directory.Exists(directory);
-    }
-
-    public void CreateDirectory(string directory)
-    {
-        Directory.CreateDirectory(directory);
-    }
-
-    public void DeleteDirectory(string directory)
-    {
-        Directory.Delete(directory, recursive: true);
-    }
-
-    public IList<string> ListFiles(string directory)
-    {
-        return Directory
-            .GetFiles(directory)
-            .Select(Path.GetFileName)
-            .OfType<string>()
-            .ToList(); 
-    }
-
-    public void Move(string sourcePath, string destinationPath)
-    {
-        Directory.Move(sourcePath, destinationPath);
-    }
-}
-
-/// <summary>
-/// This class represents an actor who is guided through the workflow
-/// as implemented in <see cref="ProcessRequestFilesWorkflow"/>.
-///
-/// This class is responsible for implementing steps that
-/// are carried out during the workflow that are specific to individual
-/// use cases e.g. collecting data about and generating Parquet report
-/// files for Public API queries.
-/// </summary>
-public interface IWorkflowActor
-{
-    string GetSourceDirectory();
-    
-    string GetReportsDirectory();
-
-    /// <summary>
-    /// Create any initial tables and state needed for collecting information
-    /// from source files.
-    /// </summary>
-    Task InitialiseDuckDb(DuckDbConnection connection);
-
-    /// <summary>
-    /// Given a source file, process it, generally by reading it into DuckDb
-    /// into tables set up at the beginning of the workflow.
-    /// </summary>
-    /// <param name="sourceFilesDirectory"></param>
-    /// <param name="connection">
-    ///     An open DuckDB connection that supports JSON reading and Parquet writing.
-    /// </param>
-    Task ProcessSourceFiles(string sourceFilesDirectory, DuckDbConnection connection);
-
-    /// <summary>
-    /// Create one or more Parquet report files based on the source files that
-    /// have been read.
-    /// </summary>
-    /// <param name="reportsFolderPathAndFilenamePrefix">
-    /// The fully-qualified folder path and filename prefix for any reports being
-    /// generated. In addition to a name suffix to identify each report being generated
-    /// here, '.parquet' will also need to be appended to any report filenames.
-    /// </param>
-    /// <param name="connection">
-    /// An open DuckDB connection that supports JSON reading and Parquet writing.
-    /// </param>
-    Task CreateParquetReports(string reportsFolderPathAndFilenamePrefix, DuckDbConnection connection);
-}
-
